refactor/rename command
