--- app/api/src/functions/oaiProxy/oaiProxy.scenarios.ts
@@ -1,6 +1,0 @@
-export const standard = defineScenario({
-  // Define the "fixture" to write into your test database here
-  // See guide: https://redwoodjs.com/docs/testing#scenarios
-})
-
-export type StandardScenario = typeof standard
