+++ vs-code-extension/src/client.ts
@@ -1,11 +1,11 @@
 import fetch from "node-fetch";
-import { OaiProxyRequest, OaiProxyResponse } from "../../app/api/types/api";
+import { EditSelectionRequest, EditSelectionResponse } from "../../app/api/types/api";
 import * as vscode from "vscode";
 
 const apiServer = vscode.workspace.getConfiguration("clippy-ai").get("apiServer");
 
-export async function createEdit(request: OaiProxyRequest) {
-  const resp = await fetch(`${apiServer}/oaiProxy`, {
+export async function createEdit(request: EditSelectionRequest) {
+  const resp = await fetch(`${apiServer}/editSelection`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
@@ -13,7 +13,7 @@ export async function createEdit(request: OaiProxyRequest) {
     body: JSON.stringify(request),
   });
 
-  const parsed = (await resp.json()) as OaiProxyResponse;
+  const parsed = (await resp.json()) as EditSelectionResponse;
   let replacement = parsed?.data?.choices?.[0]?.text;
   if (request.input?.trimEnd() === request.input) replacement = replacement?.trimEnd();
