+++ vs-code-extension/src/extension.ts
@@ -1,8 +1,10 @@
 import * as vscode from "vscode";
-import codexEdit from "./codexEdit";
+import editSelection from "./editSelection";
 
 export function activate(context: vscode.ExtensionContext) {
-  context.subscriptions.push(vscode.commands.registerCommand("clippy-ai.codexEdit", codexEdit));
+  context.subscriptions.push(
+    vscode.commands.registerCommand("clippy-ai.editSelection", editSelection)
+  );
 }
 
 export function deactivate() {}
