+++ vs-code-extension/src/recordVoiceCommand.ts
@@ -1,9 +1,9 @@
 import { spawn } from "child_process";
 import * as vscode from "vscode";
 import { readFile } from "fs/promises";
-import { OaiProxyRequest } from "../../app/api/types/api";
+import { EditSelectionRequest } from "../../app/api/types/api";
 
-const recordVoiceCommand = (): Promise<OaiProxyRequest["instruction"] | null> =>
+const recordVoiceCommand = (): Promise<EditSelectionRequest["instruction"] | null> =>
   new Promise(async (resolve, reject) => {
     try {
       const child = spawn("ffmpeg", [
@@ -20,7 +20,6 @@ const recordVoiceCommand = (): Promise<OaiProxyRequest["instruction"] | null> =>
         prompt: "Recording Clippy command from the microphone",
         placeHolder: "Press 'enter' use recorded command, or type a command here manually",
       });
-      console.log({ input });
 
       child.addListener("exit", async () => {
         if (input === undefined) {
